crewAI_web_search/requirements.txt
crewai[tools]>=0.121.0
python-dotenv>=1.0.0